Refactor output matching test utils
