+++ tests/test_tokens.py
@@ -9,43 +9,8 @@ TODO: cover more scenarios
  * streamed uploads
 
 """
-from .utils.matching import assert_output_matches, Expect
+from .utils.matching import assert_output_matches, Expect, ExpectSequence
 from .utils import http, HTTP_OK, MockEnvironment
-
-
-RAW_REQUEST = [
-    Expect.REQUEST_HEADERS,
-    Expect.BODY,
-]
-RAW_RESPONSE = [
-    Expect.RESPONSE_HEADERS,
-    Expect.BODY,
-]
-RAW_EXCHANGE = [
-    *RAW_REQUEST,
-    Expect.SEPARATOR,  # Good choice?
-    *RAW_RESPONSE,
-]
-RAW_BODY = [
-    Expect.BODY,
-]
-
-TERMINAL_REQUEST = [
-    *RAW_REQUEST,
-    Expect.SEPARATOR,
-]
-TERMINAL_RESPONSE = [
-    *RAW_RESPONSE,
-    Expect.SEPARATOR,
-]
-TERMINAL_EXCHANGE = [
-    *TERMINAL_REQUEST,
-    *TERMINAL_RESPONSE,
-]
-TERMINAL_BODY = [
-    RAW_BODY,
-    Expect.SEPARATOR
-]
 
 
 def test_headers():
@@ -60,17 +25,17 @@ def test_redirected_headers():
 
 def test_terminal_headers_and_body():
     r = http('--print=HB', '--offline', 'pie.dev', 'AAA=BBB')
-    assert_output_matches(r, TERMINAL_REQUEST)
+    assert_output_matches(r, ExpectSequence.TERMINAL_REQUEST)
 
 
 def test_terminal_request_headers_response_body(httpbin):
     r = http('--print=Hb', httpbin + '/get')
-    assert_output_matches(r, TERMINAL_REQUEST)
+    assert_output_matches(r, ExpectSequence.TERMINAL_REQUEST)
 
 
 def test_raw_request_headers_response_body(httpbin):
     r = http('--print=Hb', httpbin + '/get', env=MockEnvironment(stdout_isatty=False))
-    assert_output_matches(r, RAW_REQUEST)
+    assert_output_matches(r, ExpectSequence.RAW_REQUEST)
 
 
 def test_terminal_request_headers_response_headers(httpbin):
@@ -93,29 +58,29 @@ def test_raw_headers_and_body():
         '--print=HB', '--offline', 'pie.dev', 'AAA=BBB',
         env=MockEnvironment(stdout_isatty=False),
     )
-    assert_output_matches(r, RAW_REQUEST)
+    assert_output_matches(r, ExpectSequence.RAW_REQUEST)
 
 
 def test_raw_body():
     r = http('--print=B', '--offline', 'pie.dev', 'AAA=BBB', env=MockEnvironment(stdout_isatty=False))
-    assert_output_matches(r, RAW_BODY)
+    assert_output_matches(r, ExpectSequence.RAW_BODY)
 
 
 def test_raw_exchange(httpbin):
     r = http('--verbose', httpbin + '/post', 'a=b', env=MockEnvironment(stdout_isatty=False))
     assert HTTP_OK in r
-    assert_output_matches(r, RAW_EXCHANGE)
+    assert_output_matches(r, ExpectSequence.RAW_EXCHANGE)
 
 
 def test_terminal_exchange(httpbin):
     r = http('--verbose', httpbin + '/post', 'a=b')
     assert HTTP_OK in r
-    assert_output_matches(r, TERMINAL_EXCHANGE)
+    assert_output_matches(r, ExpectSequence.TERMINAL_EXCHANGE)
 
 
 def test_headers_multipart_body_separator():
     r = http('--print=HB', '--multipart', '--offline', 'pie.dev', 'AAA=BBB')
-    assert_output_matches(r, TERMINAL_REQUEST)
+    assert_output_matches(r, ExpectSequence.TERMINAL_REQUEST)
 
 
 def test_redirected_headers_multipart_no_separator():
@@ -123,16 +88,16 @@ def test_redirected_headers_multipart_no_separator():
         '--print=HB', '--multipart', '--offline', 'pie.dev', 'AAA=BBB',
         env=MockEnvironment(stdout_isatty=False),
     )
-    assert_output_matches(r, RAW_REQUEST)
+    assert_output_matches(r, ExpectSequence.RAW_REQUEST)
 
 
 def test_verbose_chunked(httpbin_with_chunked_support):
     r = http('--verbose', '--chunked', httpbin_with_chunked_support + '/post', 'hello=world')
     assert HTTP_OK in r
     assert 'Transfer-Encoding: chunked' in r
-    assert_output_matches(r, TERMINAL_EXCHANGE)
+    assert_output_matches(r, ExpectSequence.TERMINAL_EXCHANGE)
 
 
 def test_request_headers_response_body(httpbin):
     r = http('--print=Hb', httpbin + '/get')
-    assert_output_matches(r, TERMINAL_REQUEST)
+    assert_output_matches(r, ExpectSequence.TERMINAL_REQUEST)
