+++ tests/utils/matching/__init__.py
@@ -1,14 +1,20 @@
+"""
+Utilities for testing output composition.
+
+"""
 from typing import Iterable
 
 import pytest
 
-from .parsing import OutputMatchingError, expect_tokens, Expect
+from .parsing import OutputMatchingError, expect_tokens
+from .tokens import Expect, ExpectSequence
 
 
 __all__ = [
     'assert_output_matches',
     'assert_output_does_not_match',
     'Expect',
+    'ExpectSequence',
 ]
